Medical Store
In this project, I have implemented a medical store management system using C# language which provides a medical store owner to keep track of medicines present in his/her store. The store owner can add, delete, modify records in the file. They can also search a particular record on the basis of medicine name or generics name and chemical name, thus it will give the complete information of that medicine and chemical which includes the batch no, name, category, sales price and purchase price, also has report and chart system. This software is comprehensiveness and the simple by which it handles complex tasks easily. The medical store management system project covers important details required for designing in a Medical Store Management System with a portion of the Pharmacy and Chemical software application. The main objective of this project is to develop a software application for medical shops for performing tasks like billing, managing customer’s details, organizing stock details, Daily, weekly, Monthly, and yearly reports. The aim is to automate its existing manual system by the help of computerized equipment’s and full-fledged computer software, fulfilling their requirements, so that their valuable data/information can be stored for a longer period with easy accessing and manipulation of the same. Basically, the project describes how to manage for good performance and better services for the clients. It is providing a user friendly interface to perform all the basic operations required for a medical store thus helps in maintaining the records with ease.